chore: share the adblock use-case switcher with the sdk examples
