+++ data/git-timestamps-modified.json
@@ -555,6 +555,7 @@
   "src/pages/feature/headers.js": "2026-06-27T00:00:00.000Z",
   "src/pages/feature/proxy.js": "2026-06-27T00:00:00.000Z",
   "src/pages/feature/ttl.js": "2026-06-27T00:00:00.000Z",
+  "src/pages/features/adblock.js": "2026-08-08T00:00:00.000Z",
   "src/pages/features/antibot.js": "2026-07-29T00:00:00.000Z",
   "src/pages/features/automation.js": "2026-07-28T00:00:00.000Z",
   "src/pages/features/function.js": "2026-07-28T00:00:00.000Z",
