+++ src/components/pages/sdk/examples.js
@@ -1,9 +1,18 @@
 import React from 'react'
+import {
+  Camera,
+  FileText,
+  Filter,
+  Play,
+  Printer,
+  Search,
+  Tag
+} from 'react-feather'
 
 import Box from 'components/elements/Box'
 import Flex from 'components/elements/Flex'
 import ArrowLink from 'components/patterns/ArrowLink'
-import ExamplesSwitcher from 'components/patterns/ExamplesSwitcher'
+import UseCaseSwitcher from 'components/patterns/UseCaseSwitcher'
 import { PANELS } from 'components/pages/home/examples-panels'
 import { HOME_CONTENT_WIDTH } from 'components/pages/home/catalog'
 
@@ -15,6 +24,20 @@ import {
 } from 'components/pages/sdk/shared'
 
 import { theme } from 'theme'
+
+const PANEL_ICONS = {
+  screenshot: Camera,
+  pdf: Printer,
+  metadata: Tag,
+  markdown: FileText,
+  search: Search,
+  function: Play,
+  extract: Filter
+}
+
+const EXAMPLE_PANELS = PANELS.filter(panel => PANEL_ICONS[panel.id]).map(
+  panel => ({ ...panel, icon: PANEL_ICONS[panel.id] })
+)
 
 const Examples = () => (
   <Section id='examples'>
@@ -30,7 +53,12 @@ const Examples = () => (
         width: '100%'
       })}
     >
-      <ExamplesSwitcher panels={PANELS} visibleTabs={5} />
+      <UseCaseSwitcher
+        baseId='sdk-examples'
+        selectLabel='Choose an example'
+        tablistLabel='Examples'
+        panels={EXAMPLE_PANELS}
+      />
     </Box>
     <Flex
       css={theme({
